requirement.txt
certifi==2017.11.5
chardet==3.0.4
dj-database-url==0.4.1
Django==1.9.7
django-classy-tags==0.7.2
django-cms==3.3.1
django-embed-video==1.1.2
django-formtools==1.0
django-mptt==0.8.7
django-sekizai==0.9.0
django-tinymce==2.3.0
django-treebeard==4.0.1
djangocms-admin-style==1.2.3
djangocms-installer==0.8.10
html5lib==0.9999999
idna==2.6
Pillow==3.3.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pytz==2016.6.1
requests==2.18.4
six==1.10.0
tzlocal==1.2.2
urllib3==1.22
